Move segment and ancestry_list into a public scope for testing.
