+++ fwdpp/ts/simplification/Makefile.am
@@ -1,0 +1,4 @@
+pkgincludedir=$(prefix)/include/fwdpp/ts/simplification
+
+pkginclude_HEADERS=segment.hpp \
+				   ancestry_list.hpp
